+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- **CHANGE/ADDITION (BREAKING)** - [Kitty image protocol](https://sw.kovidgoyal.net/kitty/graphics-protocol/) support
+
 ## 0.1.0 - 8/7/2024
 
 Initial release.